Rename tutorial_controller file
